+++ pyppeteer/browser.py
@@ -10,9 +10,10 @@ from typing import Any, Awaitable, Callable, Dict, List, Optional
 
 from pyee import EventEmitter
 
-from pyppeteer.connection import Connection, CDPSession
+from pyppeteer.connection import Connection
 from pyppeteer.errors import BrowserError
 from pyppeteer.page import Page
+from pyppeteer.target import Target
 from pyppeteer.util import merge_dict
 
 
@@ -81,10 +82,17 @@ class Browser(EventEmitter):
         return browser
 
     async def _targetCreated(self, event: Dict) -> None:
-        target = Target(self, event['targetInfo'])
-        if event['targetInfo']['targetId'] in self._targets:
+        targetInfo = event['targetInfo']
+        target = Target(
+            targetInfo,
+            lambda: self._connection.createSession(targetInfo['targetId']),
+            self._ignoreHTTPSErrors,
+            self._appMode,
+            self._screenshotTaskQueue,
+        )
+        if targetInfo['targetId'] in self._targets:
             raise BrowserError('target should not exist before create.')
-        self._targets[event['targetInfo']['targetId']] = target
+        self._targets[targetInfo['targetId']] = target
         if await target._initializedPromise:
             self.emit(Browser.Events.TargetCreated, target)
 
@@ -99,7 +107,11 @@ class Browser(EventEmitter):
         target = self._targets.get(event['targetInfo']['targetId'])
         if not target:
             raise BrowserError('target should exist before targetInfoChanged')
+        previousURL = target.url
+        wasInitialized = target._isInitialized
         target._targetInfoChanged(event['targetInfo'])
+        if wasInitialized and previousURL != target.url:
+            self.emit(Browser.Events.TargetChanged, target)
 
     @property
     def wsEndpoint(self) -> str:
@@ -160,74 +172,3 @@ class Browser(EventEmitter):
 
     def _getVersion(self) -> Awaitable:
         return self._connection.send('Browser.getVersion')
-
-
-class Target(object):
-    """Browser's target class."""
-
-    def __init__(self, browser: Browser, targetInfo: Dict) -> None:
-        self._browser = browser
-        self._targetId = targetInfo.get('targetId', '')
-        self._targetInfo = targetInfo
-        self._page = None
-
-        self._initializedPromise = asyncio.get_event_loop().create_future()
-        self._isInitialized = (self._targetInfo['type'] != 'page'
-                               or self._targetInfo['url'] != '')
-        if self._isInitialized:
-            self._initializedCallback(True)
-
-    def _initializedCallback(self, bl: bool) -> None:
-        # TODO: this may cause error on page close
-        if self._initializedPromise.done():
-            self._initializedPromise = asyncio.get_event_loop().create_future()
-        self._initializedPromise.set_result(bl)
-
-    async def createCDPSession(self) -> CDPSession:
-        """Create a Chrome Devtools Protocol session attached to the target."""
-        return await self._browser._connection.createSession(self._targetId)
-
-    async def page(self) -> Optional[Page]:
-        """Get page of this target."""
-        if self._targetInfo['type'] == 'page' and self._page is None:
-            client = await self._browser._connection.createSession(
-                self._targetId)
-            new_page = await Page.create(
-                client, self,
-                self._browser._ignoreHTTPSErrors,
-                self._browser._appMode,
-                self._browser._screenshotTaskQueue,
-            )
-            self._page = new_page
-            return new_page
-        return self._page
-
-    @property
-    def url(self) -> str:
-        """Get url of this target."""
-        return self._targetInfo['url']
-
-    @property
-    def type(self) -> str:
-        """Get type of this target.
-
-        Type can be ``'page'``, ``'service_worker'``, ``'browser'``, or
-        ``'other'``.
-        """
-        _type = self._targetInfo['type']
-        if _type in ['page', 'service_worker', 'browser']:
-            return _type
-        return 'other'
-
-    def _targetInfoChanged(self, targetInfo: Dict) -> None:
-        previousURL = self._targetInfo['url']
-        self._targetInfo = targetInfo
-
-        if not self._isInitialized and (self._targetInfo['type'] != 'page' or
-                                        self._targetInfo['url'] != ''):
-            self._isInitialized = True
-            self._initializedCallback(True)
-            return
-
-        if previousURL != targetInfo['url']:
-            self._browser.emit(Browser.Events.TargetChanged, self)